move density derviative calculations to pf
