+++ pandapipes/properties/fluids.py
@@ -821,31 +821,3 @@ def create_individual_fluid(fluid_name, fluid_components,
     fluid = Fluid(fluid_name, phase, density=dens, viscosity=visc, heat_capacity=heat, molar_mass=mass,
                   der_compressibility=derc, compressibility=comp, hhv=higc, lhv=lowc)
     return fluid
-
-
-def get_derivative_density_diff(mass_fraction, density_list):
-    rho_prod = np.prod(density_list, axis=1)
-    shape = np.shape(mass_fraction)
-    loop = np.arange(0, shape[1])
-    nom = np.zeros(shape[0])
-    rho_select = np.zeros(shape)
-    for i in loop:
-        select = loop != i
-        nom += mass_fraction[:, i] * np.prod(density_list[:, select], axis=1)
-        rho_select[:, i] += np.prod(density_list[:, select], axis=1)
-    res = -rho_prod[:, np.newaxis] * rho_select * nom[:, np.newaxis] ** -2
-    return res
-
-
-def get_derivative_density_same(mass_fraction, density_list):
-    rho_prod = np.prod(density_list, axis=1)
-    shape = np.shape(mass_fraction)
-    loop = np.arange(0, shape[1])
-    nom = np.zeros(shape[0])
-    rho_select = np.zeros(shape)
-    for i in loop:
-        select = loop != i
-        nom += mass_fraction[:, i] * np.prod(density_list[:, select], axis=1)
-        rho_select[:, i] += np.prod(density_list[:, select], axis=1) * mass_fraction[:, i]
-    res = rho_prod[:, np.newaxis] * (-rho_select+nom[:, np.newaxis]) * nom[:, np.newaxis] ** -2
-    return res
